refactor after namespace moves
